Python/Lab/_PyTube/requirements.txt
git+https://github.com/nficano/pytube.git
hurry.filesize==0.9
